Rename package from pdmt5api to mt5api
Update all imports, references, and configuration to reflect the new
package name across source code, tests, documentation, and build configs.
