+++ README.md
@@ -1,5 +1,5 @@
-# pdmt5-api
+# mt5api
 
 MetaTrader 5 REST API using FastAPI
 
-[![CI/CD](https://github.com/dceoy/pdmt5-api/actions/workflows/ci.yml/badge.svg)](https://github.com/dceoy/pdmt5-api/actions/workflows/ci.yml)
+[![CI/CD](https://github.com/dceoy/mt5api/actions/workflows/ci.yml/badge.svg)](https://github.com/dceoy/mt5api/actions/workflows/ci.yml)
